Split integration tests into multiple files
